+++ src/mruby_integration/core/window.cpp
@@ -7,19 +7,6 @@
 #include "ruby/core/window.hpp"
 
 auto
-mrb_close_window(mrb_state*, mrb_value) -> mrb_value
-{
-  CloseWindow();
-  return mrb_nil_value();
-}
-
-auto
-mrb_window_ready(mrb_state*, mrb_value) -> mrb_value
-{
-  return mrb_bool_value(IsWindowReady());
-}
-
-auto
 mrb_window_state(mrb_state* mrb, mrb_value) -> mrb_value
 {
   mrb_int flag;
@@ -265,13 +252,6 @@ mrb_get_window_scale_dpi(mrb_state* mrb, mrb_value) -> mrb_value
 void
 append_core_window(mrb_state* mrb)
 {
-  mrb_define_method(
-    mrb, mrb->kernel_module, "close_window", mrb_close_window, MRB_ARGS_NONE());
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "window_ready?",
-                    mrb_window_ready,
-                    MRB_ARGS_NONE());
   mrb_define_method(mrb,
                     mrb->kernel_module,
                     "window_fullscreen?",